TeamProjectPlanner/requirements.txt
#Add the project dependencies to this file
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
sqlparse==0.5.3
filelock==3.18.0